Rename CustomPlayer to YoutubePlayer
