+++ frontend/src/app/components/CreateCover/YoutubePlayer.tsx
