+++ frontend/src/app/components/CreateCover/index.ts
@@ -1,2 +1,2 @@
-export { default as CustomPlayer } from './CustomPlayer';
+export { default as YoutubePlayer } from './YoutubePlayer';
 export { default as WaveformSetButton } from './WaveformSetButton';
